# feat: Optimize log parsing performance by using unsafe package